backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# AWS
boto3==1.34.34

# Background jobs
celery[redis]==5.3.6
redis==5.0.1

# Config & validation
pydantic==2.5.3
pydantic-settings==2.1.0

# YAML rules
pyyaml==6.0.1
